requirements.txt
pandas
django
psycopg2
django-import-export
